requirements.txt
packaging==24.2
accelerate==0.29.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
boto3==1.37.28
botocore==1.37.28
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decord==0.6.0
deepspeed==0.14.4
dill==0.3.8
distro==1.9.0
einops==0.8.1
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.3
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
open_clip_torch==2.32.0
openai==1.86.0
opencv-python==4.11.0.86
pandas==2.2.3
peft==0.4.0
pillow==11.0.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
PuLP==3.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pynvml==11.5.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==14.0.0
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.25.0
setproctitle==1.3.5
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.53.0
triton==3.0.0
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tyro==0.9.18
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.9
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.19.0